test(openai): split websocket transport regression
